Created extension methods for Swagger
Created an extension method for the Container and Http Pipeline setup for Swagger
