+++ dwCheckApi/ConfigureContainerExtenstions.cs
@@ -1,8 +1,12 @@
-﻿using dwCheckApi.Common;
+﻿using System.IO;
+using dwCheckApi.Common;
 using dwCheckApi.DAL;
+using dwCheckApi.Helpers;
 using dwCheckApi.Persistence;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.PlatformAbstractions;
+using Swashbuckle.AspNetCore.Swagger;
 
 namespace dwCheckApi
 {
@@ -44,6 +48,57 @@ namespace dwCheckApi
                         .AllowAnyMethod()
                         .AllowAnyHeader()
                         .AllowCredentials());
+            });
+        }
+
+        /// <summary>
+        /// Used to register and add the Swagger generator to the service Colelction
+        /// </summary>
+        /// <param name="serviceCollection">
+        /// The <see cref="IServiceCollection"/> which is used in the Containter
+        /// </param>
+        /// <param name="includeXmlDocumentation">
+        /// Whether or not to include XmlDocumentation (defaults to True)
+        /// </param>
+        /// <remarks>
+        /// <param name="includeXmlDocumentation"/> requries:
+        ///   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
+        ///     <DocumentationFile>bin\Debug\netcoreapp2.0\dwCheckApi.xml</DocumentationFile>
+        ///  </PropertyGroup>
+        /// for debug builds and:
+        ///   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|AnyCPU'">
+        ///     <DocumentationFile>bin\Release\netcoreapp2.0\dwCheckApi.xml</DocumentationFile>
+        ///  </PropertyGroup>
+        /// </remarks>
+        public static void AddSwagger(this IServiceCollection serviceCollection,
+            bool includeXmlDocumentation = true)
+        {
+            // Register the Swagger generator, defining one or more Swagger documents
+            serviceCollection.AddSwaggerGen(c =>
+            {
+                c.SwaggerDoc($"v{CommonHelpers.GetVersionNumber()}",
+                    new Info
+                    {
+                        Title = "dwCheckApi",
+                        Version = $"v{CommonHelpers.GetVersionNumber()}",
+                        Description = "A simple APi to get the details on Books, Characters and Series within a canon of novels",
+                        Contact = new Contact
+                        {
+                            Name = "Jamie Taylor",
+                            Email = "",
+                            Url = "https://dotnetcore.gaprogman.com"
+                        }
+                    }
+                );
+                
+                if (!includeXmlDocumentation) return;
+                // Set the comments path for the Swagger JSON and UI.
+                var basePath = PlatformServices.Default.Application.ApplicationBasePath;
+                var xmlPath = Path.Combine(basePath, "dwCheckApi.xml");
+                if (File.Exists(xmlPath))
+                {
+                    c.IncludeXmlComments(xmlPath);
+                }
             });
         }
     }
